fix(ui): streamline dashboard layout and navbar components
- Updated DashboardContent, Navbar, sessionTracker for consistent styling
- Removed redundant Navbar2 and Navbar3 components
- Adjusted overview and sessions pages for improved alignment
- Refined layout.tsx to unify authed dashboard structure
- Polished Footer and Hero sections for better visual balance
