+++ app/(authed)/dashboard/_components/Navbar.tsx
@@ -1,148 +1,148 @@
-'use client';
-import Link from "next/link";
+"use client"
+
+import React, { useState, useEffect } from 'react';
+import Link from 'next/link';
 import {
-    Home,
-    Target,
-    Clock,
-    BarChart3,
-    Download,
-    Moon,
-    User,
-    Timer,
-    LogOut
-} from "lucide-react";
-import React from 'react';
-import { useRouter, usePathname } from 'next/navigation';
+  Home,
+  Target,
+  Clock,
+  BarChart3,
+  Download,
+  User,
+  Timer,
+  LogOut,
+  Menu,
+  X,
+  BrainCircuit
+} from 'lucide-react';
+import { useRouter } from 'next/navigation';
 import { Button } from '@/components/ui/button';
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuSeparator, DropdownMenuTrigger } from '@//components/ui/dropdown-menu';
-import { useState, useEffect } from 'react';
-import { ModeToggle } from '@/components/toggle-theme'
-import {
-    onAuthStateChanged,
-    signOut,
-    User as FirebaseUser, // Rename to avoid conflict
-} from 'firebase/auth';
+import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuSeparator, DropdownMenuTrigger } from '@/components/ui/dropdown-menu';
+import { ModeToggle } from '@/components/toggle-theme';
+import { onAuthStateChanged, signOut, User as FirebaseUser } from 'firebase/auth';
 import { auth } from '@/lib/firebase';
 
-export default function Navigation() {
+export default function Navbar() {
+  const [user, setUser] = useState<FirebaseUser | null>(null);
+  const [isLoading, setIsLoading] = useState(true);
+  const [mobileOpen, setMobileOpen] = useState(false);
+  const router = useRouter();
 
-    const [user, setUser] = useState<FirebaseUser | null>(null);
-    const [isAuthReady, setIsAuthReady] = useState(false);
-    const [isLoading, setIsLoading] = useState(true);
-    const router = useRouter();
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
+      setUser(currentUser);
+      setIsLoading(false);
+    });
+    return () => unsubscribe();
+  }, []);
 
-    // NEW: Modern way to handle auth state changes with Firebase
-    // --- Auth Listener ---
-    useEffect(() => {
-        const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
-            if (currentUser) {
-                setUser(currentUser);
-                // setIsLoading(false);
-                setIsAuthReady(true); // <-- Set auth ready
-            } else {
-                setUser(null);
-            }
-            setIsLoading(false); // You already have this state
-        });
-        // Cleanup subscription on unmount
-        return () => unsubscribe();
-    }, []);
+  const handleLogout = async () => {
+    await signOut(auth);
+    router.push('/login');
+  };
 
-    const handleLogout = async () => {
-        await signOut(auth);
-        router.push("/login");
-    };
+  const navLinks = (
+    <>
+      <Link href="/dashboard/overview" onClick={() => setMobileOpen(false)} className="flex items-center gap-2 px-3 py-2 rounded hover:bg-gray-100 dark:hover:bg-gray-800">
+        <Home className="w-5 h-5" /> <span>Dashboard</span>
+      </Link>
+      <Link href="/dashboard/goals" onClick={() => setMobileOpen(false)} className="flex items-center gap-2 px-3 py-2 rounded hover:bg-gray-100 dark:hover:bg-gray-800">
+        <Target className="w-5 h-5" /> <span>Goals</span>
+      </Link>
+      <Link href="/dashboard/sessions" onClick={() => setMobileOpen(false)} className="flex items-center gap-2 px-3 py-2 rounded hover:bg-gray-100 dark:hover:bg-gray-800">
+        <Clock className="w-5 h-5" /> <span>Sessions</span>
+      </Link>
+      <Link href="/dashboard/analytics" onClick={() => setMobileOpen(false)} className="flex items-center gap-2 px-3 py-2 rounded hover:bg-gray-100 dark:hover:bg-gray-800">
+        <BarChart3 className="w-5 h-5" /> <span>Analytics</span>
+      </Link>
+      <Link href="/dashboard/export-data" onClick={() => setMobileOpen(false)} className="flex items-center gap-2 px-3 py-2 rounded hover:bg-gray-100 dark:hover:bg-gray-800">
+        <Download className="w-5 h-5" /> <span>Export</span>
+      </Link>
+    </>
+  );
 
+  return (
+    <nav className="sticky top-0 w-full border-b border-gray-200 dark:border-gray-700 bg-background z-30">
+      <div className="container mx-auto px-4">
+        <div className="flex items-center justify-between h-16">
+          {/* Left: logo */}
+          <div className="flex items-center gap-3">
+            {/* <div className="flex items-center gap-3">
+              <div className="w-10 h-10 bg-linear-to-br from-purple-600 to-purple-700 rounded-xl flex items-center justify-center">
+                <Timer className="w-6 h-6 text-white" />
+              </div>
+              <div className="hidden sm:block">
+                <h1 className="text-lg font-semibold">DeepSession</h1>
+                <p className="text-xs text-muted-foreground">Smart Session Tracking</p>
+              </div>
+            </div> */}
+            {/* Logo */}
+            <Link href="/dashboard/overview" className="flex items-center gap-2 group">
+              <div className="flex items-center justify-center w-8 h-8 rounded-lg bg-primary text-primary-foreground transition-transform group-hover:scale-105">
+                <BrainCircuit size={18} />
+              </div>
+              <span className="text-foreground font-bold text-lg tracking-tight">DeepSession</span>
+            </Link>
+          </div>
 
-    return (
-        <nav className="border-b border-white/6 bg-black/40 backdrop-blur-sm">
-            <div className="max-w-[1540px] mx-auto px-12 h-[67px] flex items-center justify-between">
-                <div className="flex items-center gap-[262px]">
-                    <Link href="/" className="flex items-center gap-3">
-                        <div className="w-[34px] h-[34px] rounded-[11px] bg-[#7F22FE] flex items-center justify-center">
-                            <Clock className="w-[21px] h-[21px] text-white" strokeWidth={1.78} />
-                        </div>
-                        {/* <span className="text-white font-normal text-[17px] leading-[26px]">
-                            DeepSession
-                        </span> */}
-                        <div>
-                            <h1 className="text-xl font-semibold text-gray-900 dark:text-white">DeepSession</h1>
-                            <p className="text-xs text-gray-500 dark:text-gray-400">Smart Session Tracking</p>
-                        </div>
-                    </Link>
+          {/* Center: desktop nav links */}
+          <div className="hidden md:flex items-center space-x-4">{navLinks}</div>
 
-                    <div className="flex items-center gap-6">
-                        <Link
-                            href="/dashboard/overview"
-                            className="flex items-center gap-2 px-4 py-2 rounded-[11px] bg-[#1E1E1E] text-white"
-                        >
-                            <Home className="w-[17px] h-[17px]" strokeWidth={1.43} />
-                            <span className="text-[17px] leading-[26px]">Dashboard</span>
-                        </Link>
+          {/* Right: actions */}
+          <div className="flex items-center gap-2">
+            <ModeToggle />
 
-                        <Link
-                            href="/dashboard/goals"
-                            className="flex items-center gap-2 px-4 py-2 rounded-[11px] text-[#A0A0A0] hover:bg-[#1E1E1E] hover:text-white transition-colors"
-                        >
-                            <Target className="w-[17px] h-[17px]" strokeWidth={1.43} />
-                            <span className="text-[17px] leading-[26px]">Goals</span>
-                        </Link>
-
-                        <Link
-                            href="/dashboard/sessions"
-                            className="flex items-center gap-2 px-4 py-2 rounded-[11px] text-[#A0A0A0] hover:bg-[#1E1E1E] hover:text-white transition-colors"
-                        >
-                            <Clock className="w-[17px] h-[17px]" strokeWidth={1.43} />
-                            <span className="text-[17px] leading-[26px]">Sessions</span>
-                        </Link>
-
-                        <Link
-                            href="/dashboard/analytics"
-                            className="flex items-center gap-2 px-4 py-2 rounded-[11px] text-[#A0A0A0] hover:bg-[#1E1E1E] hover:text-white transition-colors"
-                        >
-                            <BarChart3 className="w-[17px] h-[17px]" strokeWidth={1.43} />
-                            <span className="text-[17px] leading-[26px]">Analytics</span>
-                        </Link>
-
-                        <Link
-                            href="/dashboard/export-data"
-                            className="flex items-center gap-2 px-4 py-2 rounded-[11px] text-[#A0A0A0] hover:bg-[#1E1E1E] hover:text-white transition-colors"
-                        >
-                            <Download className="w-[17px] h-[17px]" strokeWidth={1.43} />
-                            <span className="text-[17px] leading-[26px]">Export</span>
-                        </Link>
-                    </div>
-                </div>
-
-                <div className="flex items-center gap-4">
-
-                    <ModeToggle />
-
-                    <DropdownMenu>
-                        <DropdownMenuTrigger asChild>
-                            <Button variant="ghost" size="sm">
-                                <User className="w-4 h-4 mr-2" />
-                                <span className='hidden sm:inline'> {user?.displayName?.split(' ')[0] || 'User'}</span>
-                            </Button>
-                        </DropdownMenuTrigger>
-                        <DropdownMenuContent align="end" className="w-56">
-                            <div className="px-3 py-2">
-                                <p className="text-sm text-gray-500 dark:text-gray-400">Signed in as</p>
-                                <p className="text-sm font-medium truncate flex flex-col">
-                                    <span className='inline md:hidden'> {user?.displayName?.split(' ')[0] || 'User'}</span>
-                                    <span>{user?.email}</span>
-                                </p>
-                            </div>
-                            <DropdownMenuSeparator />
-                            <DropdownMenuItem onClick={handleLogout}>
-                                <LogOut className="w-4 h-4 mr-2" />
-                                Sign out
-                            </DropdownMenuItem>
-                        </DropdownMenuContent>
-                    </DropdownMenu>
-
-                </div>
+            {/* user menu for desktop */}
+            <div className="hidden md:block">
+              <DropdownMenu>
+                <DropdownMenuTrigger asChild>
+                  <Button variant="ghost" size="sm">
+                    <User className="w-4 h-4 mr-2" />
+                    <span className="hidden md:inline">{user?.displayName?.split(' ')[0] || 'User'}</span>
+                  </Button>
+                </DropdownMenuTrigger>
+                <DropdownMenuContent align="end" className="w-56">
+                  <div className="px-3 py-2">
+                    <p className="text-sm text-muted-foreground">Signed in as</p>
+                    <p className="text-sm font-medium truncate">{user?.email}</p>
+                  </div>
+                  <DropdownMenuSeparator />
+                  <DropdownMenuItem onSelect={() => { void handleLogout(); }}>
+                    <LogOut className="w-4 h-4 mr-2" /> Sign out
+                  </DropdownMenuItem>
+                </DropdownMenuContent>
+              </DropdownMenu>
             </div>
-        </nav>
-    );
+
+            {/* Mobile menu button */}
+            <button
+              aria-label="Toggle menu"
+              className="md:hidden p-2 rounded hover:bg-gray-100 dark:hover:bg-gray-800"
+              onClick={() => setMobileOpen((s) => !s)}
+            >
+              {mobileOpen ? <X className="w-5 h-5" /> : <Menu className="w-5 h-5" />}
+            </button>
+          </div>
+        </div>
+
+        {/* Mobile drawer */}
+        <div className={`md:hidden mt-2 transition-max h-auto ${mobileOpen ? 'max-h-screen' : 'max-h-0 overflow-hidden'}`}>
+          <div className="flex flex-col gap-1 py-2">{navLinks}</div>
+          <div className="border-t border-gray-100 dark:border-gray-800 mt-2 pt-2">
+            <div className="px-3">
+              <div className="flex items-center justify-between">
+                <div>
+                  <p className="text-sm font-medium">{user?.displayName?.split(' ')[0] || 'User'}</p>
+                  <p className="text-xs text-muted-foreground truncate">{user?.email}</p>
+                </div>
+                <Button variant="ghost" size="sm" onClick={() => void handleLogout()}>
+                  <LogOut className="w-4 h-4 mr-2" /> Sign out
+                </Button>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </nav>
+  );
 }
