--- app/(authed)/dashboard/_components/Navbar2.tsx
@@ -1,147 +1,0 @@
-// app/(authed)/dashboard/_components/Navbar
-"use client"
-import React from 'react';
-import Link from "next/link";
-import {
-  Home,
-  Target,
-  Clock,
-  BarChart3,
-  Download,
-  User,
-  Timer,
-  LogOut
-} from "lucide-react";
-import { useRouter } from 'next/navigation';
-import { Button } from '@/components/ui/button';
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuSeparator, DropdownMenuTrigger } from '@//components/ui/dropdown-menu';
-import { useState, useEffect } from 'react';
-import { ModeToggle } from '@/components/toggle-theme'
-import {
-  onAuthStateChanged,
-  signOut,
-  User as FirebaseUser, // Rename to avoid conflict
-} from 'firebase/auth';
-import { auth } from '@/lib/firebase';
-
-export default function Navbar() {
-  const [isDark, setIsDark] = useState(false);
-  const [isAuthReady, setIsAuthReady] = useState(false);
-  const [isLoading, setIsLoading] = useState(true);
-  const [activeTab, setActiveTab] = useState("dashboard");
-  const [user, setUser] = useState<FirebaseUser | null>(null);
-  const router = useRouter();
-
-  // NEW: Modern way to handle auth state changes with Firebase
-  // --- Auth Listener ---
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
-      if (currentUser) {
-        setUser(currentUser);
-        // setIsLoading(false);
-        setIsAuthReady(true); // <-- Set auth ready
-      } else {
-        setUser(null);
-      }
-      setIsLoading(false); // You already have this state
-    });
-    // Cleanup subscription on unmount
-    return () => unsubscribe();
-  }, []);
-
-  const handleLogout = async () => {
-    await signOut(auth);
-    router.push("/login");
-  };
-
-  return (
-    <nav className="w-screen border-b border-gray-200 dark:border-gray-700 glass">
-      <div className="container mx-auto px-4">
-        <div className="flex items-center justify-between h-16">
-          {/* Logo */}
-          <div className="flex items-center space-x-3">
-            <div className="w-10 h-10 bg-linear-to-br from-purple-600 to-purple-700 rounded-xl flex items-center justify-center">
-              <Timer className="w-6 h-6 text-white" />
-            </div>
-            <div>
-              <h1 className="text-xl font-semibold text-gray-900 dark:text-white">DeepSession</h1>
-              <p className="text-xs text-gray-500 dark:text-gray-400">Smart Session Tracking</p>
-            </div>
-          </div>
-
-          {/* Dashbord tabs */}
-          <div className="hidden md:flex space-x-8">
-            <Link
-              href="/dashboard/overview"
-              className="flex items-center gap-2 px-4 py-2 rounded-[11px] bg-[#1E1E1E] text-white"
-            >
-              <Home className="w-[17px] h-[17px]" strokeWidth={1.43} />
-              <span className="text-[17px] leading-[26px]">Dashboard</span>
-            </Link>
-
-            <Link
-              href="/dashboard/goals"
-              className="flex items-center gap-2 px-4 py-2 rounded-[11px] text-[#A0A0A0] hover:bg-[#1E1E1E] hover:text-white transition-colors"
-            >
-              <Target className="w-[17px] h-[17px]" strokeWidth={1.43} />
-              <span className="text-[17px] leading-[26px]">Goals</span>
-            </Link>
-
-            <Link
-              href="/dashboard/sessions"
-              className="flex items-center gap-2 px-4 py-2 rounded-[11px] text-[#A0A0A0] hover:bg-[#1E1E1E] hover:text-white transition-colors"
-            >
-              <Clock className="w-[17px] h-[17px]" strokeWidth={1.43} />
-              <span className="text-[17px] leading-[26px]">Sessions</span>
-            </Link>
-
-            <Link
-              href="/dashboard/analytics"
-              className="flex items-center gap-2 px-4 py-2 rounded-[11px] text-[#A0A0A0] hover:bg-[#1E1E1E] hover:text-white transition-colors"
-            >
-              <BarChart3 className="w-[17px] h-[17px]" strokeWidth={1.43} />
-              <span className="text-[17px] leading-[26px]">Analytics</span>
-            </Link>
-
-            <Link
-              href="/dashboard/export-data"
-              className="flex items-center gap-2 px-4 py-2 rounded-[11px] text-[#A0A0A0] hover:bg-[#1E1E1E] hover:text-white transition-colors"
-            >
-              <Download className="w-[17px] h-[17px]" strokeWidth={1.43} />
-              <span className="text-[17px] leading-[26px]">Export</span>
-            </Link>
-          </div>
-
-          {/* User Menu */}
-          <div className="flex items-center space-x-4">
-
-            <ModeToggle />
-
-            <DropdownMenu>
-              <DropdownMenuTrigger asChild>
-                <Button variant="ghost" size="sm">
-                  <User className="w-4 h-4 mr-2" />
-                  <span className='hidden sm:inline'> {user?.displayName?.split(' ')[0] || 'User'}</span>
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="end" className="w-56">
-                <div className="px-3 py-2">
-                  <p className="text-sm text-gray-500 dark:text-gray-400">Signed in as</p>
-                  <p className="text-sm font-medium truncate flex flex-col">
-                    <span className='inline md:hidden'> {user?.displayName?.split(' ')[0] || 'User'}</span>
-                    <span>{user?.email}</span>
-                  </p>
-                </div>
-                <DropdownMenuSeparator />
-                <DropdownMenuItem onClick={handleLogout}>
-                  <LogOut className="w-4 h-4 mr-2" />
-                  Sign out
-                </DropdownMenuItem>
-              </DropdownMenuContent>
-            </DropdownMenu>
-          </div>
-        </div>
-      </div>
-    </nav>
-  );
-}
