+++ app/(authed)/dashboard/overview/page.tsx
@@ -15,10 +15,6 @@ export default function Dashboard() {
     <div className="space-y-6">
       {/* SSR Shell */}
       <div className="text-center space-y-4">
-        <h1 className="text-3xl font-medium">FocusFlow</h1>
-        <p className="text-muted-foreground">
-          Track your productivity and build better work habits
-        </p>
         <SessionTracker />
       </div>
